requirements.txt
# Data & Analysis
yfinance==0.2.32
pandas==2.2.0
numpy==1.26.4
ta==0.10.2

# Machine Learning
tensorflow==2.15.0
scikit-learn==1.4.1

# LLM & APIs
openai==1.3.6
requests==2.31.0

# News & Web Scraping
feedparser==6.0.10
beautifulsoup4==4.12.2
lxml==4.9.3

# Database
sqlite3

# CLI & Configuration
click==8.1.7
pydantic==2.5.0
pyyaml==6.0.1

# Scheduling
APScheduler==3.10.4

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Utilities
python-dotenv==1.0.0
